the first task errors fixing

+++ main.h
@@ -2,7 +2,7 @@
 #define MAIN_H
 int printc(char c, int *len);
 int _printf(const char *format, ...);
-void print_number(int n, int *len);
+void print_number(long n, int *len);
 char *custum_string(char *s);
 char *converttobase(unsigned int n, int base);
 void print_rev(char *s, int *len);
